Refact: Change Modal more componentize
